+++ src/access/universal.ts
@@ -452,4 +452,4 @@ export async function setPublicAccess(
   return null;
 }
 
-export { getAccessFor } from "./accessFor";
+export { getAccessFor } from "./for";
